Removed duplicate files, refactored existing code & made terminal logs look better
